chore(refactor): making history structures unite, arranging structures
