+++ src/handlers/mod.rs
@@ -1,5 +1,4 @@
 use {
-    crate::providers::zerion::ZerionTransactionTransfer,
     serde::{Deserialize, Serialize},
     wc::metrics::TaskMetrics,
 };
@@ -25,62 +24,4 @@ pub struct RpcQueryParams {
 #[derive(Serialize)]
 pub struct SuccessResponse {
     status: String,
-}
-
-#[derive(Debug, Deserialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct PortfolioQueryParams {
-    pub project_id: String,
-    pub currency: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct PortfolioResponseBody {
-    pub data: Vec<PortfolioPosition>,
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct PortfolioPosition {
-    pub id: String,
-    pub name: String,
-    pub symbol: String,
-}
-
-// TODO: https://developers.zerion.io/reference/listwallettransactions
-#[derive(Debug, Deserialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct HistoryQueryParams {
-    pub currency: Option<String>,
-    pub project_id: String,
-    pub cursor: Option<String>,
-    pub onramp: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct HistoryResponseBody {
-    pub data: Vec<HistoryTransaction>,
-    pub next: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct HistoryTransaction {
-    pub id: String,
-    pub metadata: HistoryTransactionMetadata,
-    pub transfers: Option<Vec<ZerionTransactionTransfer>>,
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Hash, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct HistoryTransactionMetadata {
-    pub operation_type: String,
-    pub hash: String,
-    pub mined_at: String,
-    pub sent_from: String,
-    pub sent_to: String,
-    pub status: String,
-    pub nonce: usize,
 }
